Implement GUI system, Part 1

+++ Includes/CubbyTower/Components/Position.hpp
@@ -16,8 +16,8 @@ namespace CubbyTower
 //!
 struct Position 
 {
-    double x;
-    double y;
+    float x;
+    float y;
 };
 }  // namespace CubbyTower
